+++ src/utils/utils.js
@@ -1,0 +1,5 @@
+const getRadomNumber = (start, end) =>
+  Math.round(Math.random() * (end - start) + start);
+
+
+export {getRadomNumber };
